refactor: split high-complexity files into focused siblings
Split three files whose per-file complexity exceeded Codacy's threshold
into focused sibling files. Pure code moves — no logic changes, no
signature changes; package coverage is unchanged (dflint 63.6%, ddc
65.5%, dockerbuild 81.0%).

- internal/dflint/lint.go (56 -> 26): built-in check rules and the RUN
  block parser move to checks.go.
- internal/ddc/ddc.go (53 -> 5): path resolution -> paths.go, size
  walking -> size.go, clean/prune ops -> ops.go; ddc.go keeps the mode
  constants, validation, and env/Zen path helpers.
- internal/dockerbuild/game.go (69 -> 13): resolvers -> game_resolver.go,
  build-script generation -> game_script.go, DDC mount args ->
  game_ddc.go, container execution -> game_container.go; game.go keeps
  the option/builder types and the Build/BuildClient entry points.

File naming mirrors the existing *_test.go split in each package.
