+++ internal/ddc/ddc.go
@@ -1,15 +1,8 @@
 package ddc
 
 import (
-	"errors"
 	"fmt"
-	"io/fs"
-	"os"
-	"os/user"
-	"path/filepath"
-	"runtime"
 	"strings"
-	"time"
 )
 
 // DDC mode constants. Use these instead of raw string comparisons.
@@ -64,200 +57,3 @@ func EnvOverride(path string) string {
 // UE uses the Zen Store as its default local DDC backend from 5.4 onward, so
 // mounting a host directory here persists the cook DDC across --rm container runs.
 const ZenContainerPath = "/home/ue/.config/Epic/UnrealEngine/Common/Zen/Data"
-
-// resolveHome returns the user's home directory. It prefers os.UserHomeDir()
-// ($HOME / %USERPROFILE%), but that fails when the environment is stripped of
-// HOME — which happens under SSM Run Command, some CI runners, and bare service
-// contexts. In that case it falls back to the os/user database, then to /root
-// for the common root-in-container case, so DDC path resolution doesn't hard-fail.
-func resolveHome() (string, error) {
-	if home, err := os.UserHomeDir(); err == nil && home != "" {
-		return home, nil
-	}
-	if u, err := user.Current(); err == nil && u.HomeDir != "" {
-		return u.HomeDir, nil
-	}
-	if runtime.GOOS != "windows" {
-		// Last-resort default for HOME-less *nix contexts (e.g. SSM as root).
-		return "/root", nil
-	}
-	return "", fmt.Errorf("resolving home directory: HOME is unset and no fallback is available; set HOME or configure ddc.zenPath/ddc.localPath explicitly")
-}
-
-// DefaultZenPath returns the default ZenStore host directory: ~/.ludus/zen
-func DefaultZenPath() (string, error) {
-	home, err := resolveHome()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(home, ".ludus", "zen"), nil
-}
-
-// ResolveZenPath returns the override path if non-empty, otherwise DefaultZenPath.
-func ResolveZenPath(override string) (string, error) {
-	if override != "" {
-		if !filepath.IsAbs(override) {
-			return "", fmt.Errorf("DDC zen path must be absolute (got %q)", override)
-		}
-		return override, nil
-	}
-	return DefaultZenPath()
-}
-
-// DefaultPath returns the default DDC directory path: ~/.ludus/ddc
-func DefaultPath() (string, error) {
-	home, err := resolveHome()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(home, ".ludus", "ddc"), nil
-}
-
-// ResolvePath returns the override path if non-empty, otherwise returns DefaultPath.
-// Returns an error if the override is a relative path.
-func ResolvePath(override string) (string, error) {
-	if override != "" {
-		if !filepath.IsAbs(override) {
-			return "", fmt.Errorf("DDC path must be absolute (got %q); use a full path like /home/user/.ludus/ddc", override)
-		}
-		return override, nil
-	}
-	return DefaultPath()
-}
-
-// DirSize returns the total bytes of all files under dir.
-// Returns 0 without error if dir doesn't exist.
-// Files that vanish mid-walk (e.g. evicted by the engine) are skipped silently.
-func DirSize(dir string) (int64, error) {
-	if dir == "" {
-		return 0, nil
-	}
-	var total int64
-	err := filepath.WalkDir(dir, func(_ string, d fs.DirEntry, err error) error {
-		if err != nil {
-			if errors.Is(err, fs.ErrNotExist) {
-				return nil // file/dir vanished between readdir and stat
-			}
-			return err
-		}
-		if d.IsDir() {
-			return nil
-		}
-		info, err := d.Info()
-		if err != nil {
-			if errors.Is(err, fs.ErrNotExist) {
-				return nil // file vanished between WalkDir discovering it and Info()
-			}
-			return err
-		}
-		total += info.Size()
-		return nil
-	})
-	if errors.Is(err, fs.ErrNotExist) {
-		return 0, nil
-	}
-	return total, err
-}
-
-// Clean removes all contents of dir (not dir itself) and returns bytes freed.
-// Returns 0 without error if dir doesn't exist.
-func Clean(dir string) (int64, error) {
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		if errors.Is(err, fs.ErrNotExist) {
-			return 0, nil
-		}
-		return 0, fmt.Errorf("reading DDC directory: %w", err)
-	}
-	var freed int64
-	for _, entry := range entries {
-		path := filepath.Join(dir, entry.Name())
-		freed += entrySize(path, entry)
-		if err := os.RemoveAll(path); err != nil {
-			return freed, fmt.Errorf("removing %s: %w", entry.Name(), err)
-		}
-	}
-	return freed, nil
-}
-
-// entrySize returns the size of an entry. For files it returns the file size;
-// for directories it walks recursively to sum all file sizes.
-func entrySize(path string, d fs.DirEntry) int64 {
-	if !d.IsDir() {
-		info, err := d.Info()
-		if err != nil {
-			return 0
-		}
-		return info.Size()
-	}
-	var total int64
-	_ = filepath.WalkDir(path, func(_ string, wd fs.DirEntry, err error) error {
-		if err != nil || wd.IsDir() {
-			return err
-		}
-		info, err := wd.Info()
-		if err != nil {
-			return nil
-		}
-		total += info.Size()
-		return nil
-	})
-	return total
-}
-
-// Prune removes files older than maxAgeDays and returns bytes freed.
-// Returns 0 without error if dir doesn't exist.
-// maxAgeDays must be at least 1 to prevent accidental deletion of all files.
-func Prune(dir string, maxAgeDays int) (int64, error) {
-	if maxAgeDays < 1 {
-		return 0, fmt.Errorf("max age must be at least 1 day (got %d)", maxAgeDays)
-	}
-	cutoff := time.Now().Add(-time.Duration(maxAgeDays) * 24 * time.Hour)
-	freed, err := removeOldFiles(dir, cutoff)
-	if errors.Is(err, fs.ErrNotExist) {
-		return 0, nil
-	}
-	return freed, err
-}
-
-// removeOldFiles walks dir and removes files with modtime before cutoff.
-func removeOldFiles(dir string, cutoff time.Time) (int64, error) {
-	var freed int64
-	err := filepath.WalkDir(dir, func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			if errors.Is(err, fs.ErrNotExist) {
-				return nil // entry vanished between readdir and visit
-			}
-			return err
-		}
-		if d.IsDir() {
-			return nil
-		}
-		return pruneIfOld(path, d, cutoff, &freed)
-	})
-	return freed, err
-}
-
-// pruneIfOld removes a single file if its modtime is before cutoff,
-// accumulating freed bytes. Files that disappear mid-walk are skipped silently.
-func pruneIfOld(path string, d fs.DirEntry, cutoff time.Time, freed *int64) error {
-	info, err := d.Info()
-	if err != nil {
-		if errors.Is(err, fs.ErrNotExist) {
-			return nil // file vanished between WalkDir discovering it and Info()
-		}
-		return err
-	}
-	if !info.ModTime().Before(cutoff) {
-		return nil
-	}
-	size := info.Size()
-	if err := os.Remove(path); err != nil {
-		if errors.Is(err, fs.ErrNotExist) {
-			return nil // already removed by another process
-		}
-		return err
-	}
-	*freed += size
-	return nil
-}
